added demo page, changed some refactoring files
