+++ src/index.js
@@ -1,67 +1,9 @@
-let content = {}
-let ftns = document.querySelectorAll("t-ftn"); //selects all footnotes
-
-//options
-let options = document.querySelector("t-options");
-var hover = options.dataset.hover === 'true';
-var unhover = options.dataset.unhoverdelete === 'true';
-let hoverDelay = options.dataset.hoverdelay === undefined ? 0 : options.dataset.hoverdelay;
-
-var slider = document.getElementById("myRange");
-slider.oninput = function() {
-    hoverDelay = this.value;
-}
+import {removeAllFootnotes, initializeFootnotes, toggleHover, toggleUnhover} from './actions.js'
+import {distance} from './utils.js'
 
 //removes all footnotes except the one that is opened
 window.addEventListener("click", (e) => removeAllFootnotes(e.target));
 initializeFootnotes();
-
-function toggleHover(){
-    hover = !hover;
-    initializeFootnotes();
-}
-
-function toggleUnhover(){
-    unhover = !unhover;
-    initializeFootnotes();
-}
-
-function initializeFootnotes(){
-
-    for(i of ftns){
-        content[i] = i.innerHTML;
-        i.removeEventListener("mouseenter", addFootnote);
-        i.removeEventListener("click", removeFootnote);
-        i.removeEventListener("click", toggleFootnote);
-        i.removeEventListener("mouseout", removeFootnoteWithTimer);
-    
-        // hover/click option hub (more options to be added)
-        if(hover){
-            i.addEventListener("mouseenter", addFootnote);
-            i.addEventListener("click", removeFootnote);
-        }else{
-            i.addEventListener("click", toggleFootnote);  
-        }
-    
-        if(unhover){
-            i.addEventListener("mouseout", removeFootnoteWithTimer);
-        }
-        
-    }
-}
-
-// calculates distance for the arrow
-function distance(tdiv, bdiv) {
-    const trect = tdiv.getBoundingClientRect();
-    const brect = bdiv.getBoundingClientRect();
-    const bleftx = brect.left;
-    const blefty = brect.top;
-    const tcenterx = trect.left + trect.width / 2;
-    const tbottomy = trect.bottom;
-    const distance = Math.hypot(bleftx - tcenterx, tbottomy - blefty);
-
-    return distance;
-}
 
 // Create a class for the element
 class Footnote extends HTMLElement {
@@ -120,42 +62,3 @@ class Footnote extends HTMLElement {
   }
   
 customElements.define("t-ftn", Footnote);
-  
-
-// toggles visibility when the note is clicked
-function toggleFootnote(){
-    removeAllFootnotes(this);
-    var vis = this.shadowRoot.children[1].style.visibility;
-    if(vis == 'hidden'){
-        this.shadowRoot.children[1].style.visibility = 'visible';
-    }else{
-        this.shadowRoot.children[1].style.visibility = 'hidden';
-    }
-}
-
-// adds visibility to clicked footnote
-function addFootnote(){
-    removeAllFootnotes();
-    this.shadowRoot.children[1].style.visibility = 'visible';
-}
-
-// removes visibility from clicked footnote
-function removeFootnote(){
-    this.shadowRoot.children[1].style.visibility = 'hidden';
-}
-
-async function removeFootnoteWithTimer(){
-    setTimeout(() => {
-        this.shadowRoot.children[1].style.visibility = 'hidden';
-    }, hoverDelay);
-}
-
-// removes visibility from all footnotes except for current note (if there is one)
-// curr = current note
-function removeAllFootnotes(curr){
-    document.querySelectorAll('t-ftn').forEach(ftn => {
-        if(ftn !== curr){
-            ftn.shadowRoot.children[1].style.visibility = 'hidden';
-        }
-    })
-}
